requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
mediapipe
numpy
joblib
tensorflow-cpu
Pillow